feat: Refactor and fix task creation form
This commit resolves a critical issue where the task creation form was non-functional due to a combination of missing fields and a duplicated, incorrect frontend component.

The following changes were made:

- **Component Refactoring:**
  - Consolidated all task creation logic into the `TaskManager.tsx` component, which is the correct one used in the `/admin` page.
  - Deleted the redundant and unused `TaskCreationPage.tsx` file to eliminate confusion.
  - Removed the now-defunct `/tasks/create` route from the main router in `App.tsx`.

- **Form Enhancements:**
  - Added all previously missing fields to the `TaskManager` form:
    - "Origin Office" dropdown.
    - "Status" dropdown.
    - Date and time pickers.
    - Checkboxes for user roles ("Responsible", "Executor", "Solicitante").
  - Updated the form's submission logic to send the complete and correctly structured payload to the backend API.

- **Backend Support:**
  - Created a new endpoint `/api/v1/offices` to provide the list of offices required by the form.
  - Registered the new router in `main.py`.

- **Critical Bug Fixes:**
  - Fixed a fatal `SyntaxError` in the backend caused by a leftover merge conflict marker.
  - Created the missing `frontend/src/lib/utils.ts` file, which was causing the entire frontend application to crash on startup.
  - Added a `.env` file with placeholder values to allow the application to start correctly in development environments.
