+++ frontend/src/App.tsx
@@ -6,7 +6,6 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Index from "./pages/Index";
 import NotFound from "./pages/NotFound";
 import AdminPage from "./pages/AdminPage";
-import TaskCreationPage from "./pages/TaskCreationPage";
 
 const queryClient = new QueryClient();
 
@@ -19,7 +18,6 @@ const App = () => (
         <Routes>
           <Route path="/" element={<Index />} />
           <Route path="/admin" element={<AdminPage />} />
-          <Route path="/tasks/create" element={<TaskCreationPage />} />
           <Route path="*" element={<NotFound />} />
         </Routes>
       </BrowserRouter>
